+++ benchmark/_run_benchmarks.R
@@ -1,0 +1,2 @@
+source("benchmark/_init.R")
+sapply(list.files("benchmark", "^bench\\-.*\\.R$", full.names = TRUE), source)