+++ tests/testthat/test-VisualizeMatching.R
@@ -18,19 +18,18 @@ test_that("VisualizeMatching() works", {
   skip_if(packageVersion("graphics") < "4.3")
   skip_if(packageVersion("vdiffr") < "1.0")
   
-  TestVM <- function() {
+  vdiffr::expect_doppelganger("Test VM", TestVM <- function() {
     VisualizeMatching(MutualClusteringInfo, tree1, tree2, 
                       setPar = TRUE, precision = 3, matchZeros = FALSE,
-                      Plot = plot.phylo)
-  }
-  vdiffr::expect_doppelganger("Test VM", TestVM)
+                      Plot = plot.phylo) -> x
+    expect_equal(sum(attr(x, "matchedScores")), x[[1]])
+  })
   
-  TestVMr <- function() {
+  vdiffr::expect_doppelganger("Test VMr", function() {
     VisualizeMatching(MutualClusteringInfo, tree1, tree2r,
                       setPar = TRUE, precision = 3, matchZeros = TRUE, 
                       Plot = plot.phylo, cex = 1.5)
-  }
-  vdiffr::expect_doppelganger("Test VMr", TestVMr)
+  })
   
   vdiffr::expect_doppelganger("Visualize MCI matching", function() {
     par(mfrow = c(2, 2), mar = rep(0.1, 4), cex = 1.5)
@@ -54,6 +53,7 @@ test_that("VisualizeMatching() works", {
     tree2 <- ape::read.tree(text="((1, 2), ((3, (4, (5, 9))), (6, (7, 8))));")
     VisualizeMatching(RobinsonFouldsMatching, tree1, tree2,
                       setPar = FALSE, precision = 3,
+                      edge.cex = 1,
                       Plot = TreeDistPlot,
                       matchZeros = TRUE,
                       leaveRoom = FALSE)
@@ -61,6 +61,26 @@ test_that("VisualizeMatching() works", {
                       setPar = FALSE, precision = 3,
                       Plot = TreeDistPlot,
                       matchZeros = FALSE,
+                      leaveRoom = FALSE)
+  })
+  
+  vdiffr::expect_doppelganger("Hidden edge labels", function() {
+    par(mfrow = c(2, 2), mar = rep(0.1, 4), cex = 1.2)
+    tree1 <- ape::read.tree(text="((1, 2), ((6, (7, 8)), (3, 4, (5, 9))));")
+    tree2 <- ape::read.tree(text="((1, 2), ((3, (4, (5, 9))), (6, (7, 8))));")
+    VisualizeMatching(RobinsonFouldsMatching, tree1, tree2,
+                      setPar = FALSE, precision = 3,
+                      edge.cex = 0,
+                      value.cex = 2,
+                      Plot = TreeDistPlot,
+                      matchZeros = TRUE,
+                      leaveRoom = FALSE)
+    VisualizeMatching(RobinsonFouldsMatching, tree1, tree2,
+                      setPar = FALSE, precision = 3,
+                      edge.cex = 2,
+                      value.cex = FALSE,
+                      Plot = TreeDistPlot,
+                      matchZeros = TRUE,
                       leaveRoom = FALSE)
   })
   
@@ -119,6 +139,7 @@ test_that("VisualizeMatching() handles unrooted trees", {
     tree1 <- UnrootTree(BalancedTree(1:5))
     tree2 <- UnrootTree(PectinateTree(1:5))
     VisualizeMatching(RobinsonFouldsMatching, tree1, tree2,
+                      edge.frame = "n",
                       setPar = FALSE,
                       Plot = TreeDistPlot)
   })